==> 3. Introduction to Data Analysis/Lesson 3 THE Data Analysis Process/15 - Assessing and Building Intuition Quiz.ipynb <==
# Assessing and Building Intuition Quiz
Use the space below to explore `census_income_data.csv` to answer the quiz questions below.



# Assessing and Building Intuition Quiz

## Using the functions you learned so far, 
### explore census_income_data.csv in the Jupyter notebook below to answer quiz questions below the notebook about these characteristics of the dataset:

* number of rows
* number of columns
* data types of several columns
* columns with missing values
* number of unique values of education
* mean age
* 75th percentile of hours-per-week

[# This data was originally taken from here.] (https://archive.ics.uci.edu/ml/datasets/Census+Income) 



# Import pandas
import pandas as pd
import numpy as np

# Load census income data
census_income=pd.read_csv("census_income_data.csv")
census_income

# Work to answer the quiz questions
census_income.dtypes

type(census_income["workclass"][0])

type(census_income["education"][0])

type(census_income["relationship"][0])

type(census_income["income"][0])

census_income.describe()

census_income.info()

census_income.nunique()
