==> 3. Introduction to Data Analysis/Lesson 3 THE Data Analysis Process/18 - cleaning_practice.ipynb <==
# Cleaning Practice
Let's first practice handling missing values and duplicate data using the `cancer_data_means.csv` file, which you created and saved in the "Assessing and Building Intuition" notebook a few pages back. If you created this CSV file in that notebook, it should still be available in this workspace for you to load into the notebook here.

# import pandas and load cancer data
import pandas as pd
import numpy as np

df=pd.read_csv("cancer_data_means.csv")
df

df.info()

mean =df["texture_mean"].mean()
print(mean)

# check which columns have missing values with info()
#df.info()
#type(df["diagnosis"])
dataTypeObj = df.dtypes["perimeter_mean"]
print(dataTypeObj)

 def mean_return(colomn):
    mean = df[colomn].mean()
    return mean

# use means to fill in missing values
###############################################################################################
""" 
df["texture_mean"]=df["texture_mean"].fillna(mean_return("texture_mean"))
df["smoothness_mean"]=df["smoothness_mean"].fillna(mean_return("smoothness_mean"))
df["symmetry_mean"]=df["symmetry_mean"].fillna(mean_return("symmetry_mean")) 
"""
###############################################################################################
def mean_return(colomn):
    mean = df[colomn].mean()
    return mean
#print(mean_return("texture_mean"))   

for i in df.columns:
    dataTypeObj = df.dtypes[i]
    if dataTypeObj == np.float64 :
        #print(type(i))
        df[i]=df[i].fillna(mean_return(i)) 
        # 

df
# confirm your correction with info()


df.info()

# check for duplicates in the data
sum(df.duplicated())

# drop duplicates
df.drop_duplicates(inplace=True)

# confirm correction by rechecking for duplicates in the data
sum(df.duplicated())

## Renaming Columns
Since we also previously changed our dataset to only include means of tumor features, the "_mean" at the end of each feature seems unnecessary. It just takes extra time to type in our analysis later. Let's come up with a list of new labels to assign to our columns.

# remove "_mean" from column names
new_labels = []
for col in df.columns:
    if '_mean' in col:
        new_labels.append(col[:-5])  # exclude last 6 characters
    else:
        new_labels.append(col)

# new labels for our columns
new_labels

# assign new labels to columns in dataframe
df.columns = new_labels

# display first few rows of dataframe to confirm changes
df.head()

# save this for later
df.to_csv('cancer_data_edited.csv', index=False)

df=pd.read_csv("cancer_data_edited.csv")
df
