==> 3. Introduction to Data Analysis/Lesson 3 THE Data Analysis Process/21 - Exploring Data with Visuals Quiz-visuals_quiz.ipynb <==
# Exploring Data with Visuals Quiz
Use the space below to explore `powerplant_data_edited.csv` to answer the quiz questions below. Remember, you created this dataset at the bottom of the notebook in the "Reading CSV Files" section earlier in the lesson! Check your work with the solutions on that page if you are having trouble getting this dataset.

# import and load data


# plot relationship between temperature and electrical output


# plot distribution of humidity


# plot box plots for each variable
