==> 3. Introduction to Data Analysis/Lesson 3 THE Data Analysis Process/23 - Drawing Conclusions Example.ipynb <==
import numpy as np
import pandas as pd

df_cancer = pd.read_csv("cancer_data_edited.csv")
df_cancer.head(10)

mask = df_cancer["diagnosis"] =="M"
print(mask)

df_m =  df_cancer [ df_cancer["diagnosis"]=="M" ]
df_m.head(10)

df_m = df_cancer[mask]
df_m

df_m["area"].describe()

df_b = df_cancer [ df_cancer["diagnosis"]=="B" ]
df_b

df_b["area"].describe()

## : .4f == the first 4 DIGIT FROM THE RETURNED FLOAT
print( "The main of M {0:.4f} , while the maen of B is {1:.4f} ".format( df_m["area"].mean()  ,  df_b["area"].mean() )  )
