==> 3. Introduction to Data Analysis/Lesson 3 THE Data Analysis Process/26 - Communicating Results Example.ipynb <==
import numpy as np
import pandas as pd

%matplotlib

df_census=pd.read_csv("census_income_data.csv")


df_above = df_census [ df_census["income"] == '>50K' ]
df_above
df_belue = df_census [ df_census["income"] == '<50K' ]
df_belue


df_above["education"].value_counts().plot(kind="bar");

df_belue["education"].value_counts().plot(kind="bar");
