==> 3. Introduction to Data Analysis/Lesson 3 THE Data Analysis Process/27 - Communicating Results Practice.py <==
# -*- coding: utf-8 -*-
"""
Created on Sun Mar  7 10:47:48 2021

@author: KING
"""

import numpy as np
import  pandas as pd
import  matplotlib as plt

df_stores = pd.read_csv("store_data.csv")
print(df_stores)
